code/requirements.txt
aioice==0.10.1
aiortc==1.12.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
av==14.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
cryptography==44.0.3
dnspython==2.7.0
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.5.0
gitdb==4.0.12
GitPython==3.1.44
google-crc32c==1.7.1
h11==0.16.0
idna==3.10
ifaddr==0.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mpmath==1.3.0
narwhals==1.39.0
networkx==3.4.2
numpy==2.2.5
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.2.1
protobuf==6.31.0
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
pylibsrtp==0.12.0
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
streamlit==1.45.1
streamlit-webrtc==0.62.4
sympy==1.14.0
tenacity==9.1.2
toml==0.10.2
torch==2.7.0
torchvision==0.22.0
tornado==6.4.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
watchdog==6.0.0
